[Encoder] expand the mb_idx range so as to support frame size > 32767

+++ codec/encoder/core/inc/encoder_context.h
@@ -204,7 +204,7 @@ typedef struct TagWelsEncCtx {
   bool						bRefOfCurTidIsLtr[MAX_DEPENDENCY_LAYER][MAX_TEMPORAL_LEVEL];
   uint16_t        uiIdrPicId;		// IDR picture id: [0, 65535], this one is used for LTR
 
-  int16_t						iMaxSliceCount;// maximal count number of slices for all layers observation
+  int32_t						iMaxSliceCount;// maximal count number of slices for all layers observation
   int16_t						iActiveThreadsNum;	// number of threads active so far
 
   /*
